requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-summernote==0.8.10.0
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
Unidecode==1.0.22
urllib3==1.24
